requirements.txt
streamlit
google-generativeai
transformers>=4.36.2
torch
pandas
sentencepiece==0.1.99